apps/api/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Environment and configuration
python-dotenv==1.0.1
pydantic-settings==2.1.0

# CORS and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File handling
boto3==1.34.34
python-magic==0.4.27
Pillow==10.2.0
ezdxf==1.1.4

# Payments
stripe==8.4.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
